+++ lib/honeybadger/sender.rb
@@ -5,6 +5,9 @@ module Honeybadger
                    Errno::EINVAL,
                    Errno::ECONNRESET,
                    EOFError,
+                   Net::HTTPBadResponse,
+                   Net::HTTPHeaderSyntaxError,
+                   Net::ProtocolError,
                    Errno::ECONNREFUSED].freeze
 
     def initialize(options = {})
@@ -41,23 +44,19 @@ module Honeybadger
       data = notice.is_a?(String) ? notice : notice.to_json
 
       response = begin
-                   client.post do |p|
-                     p.url NOTICES_URI
-                     p.body = data
-                     p.headers.merge!({ 'X-API-Key' => api_key})
-                   end
+                   http_connection.post(url.path, data, http_headers({'X-API-Key' => api_key}))
                  rescue *HTTP_ERRORS => e
                    log(:error, "Unable to contact the Honeybadger server. HTTP Error=#{e}")
                    nil
                  end
 
-      if response.success?
+      if Net::HTTPSuccess === response
         log(Honeybadger.configuration.debug ? :info : :debug, "Success: #{response.class}", response, data)
         JSON.parse(response.body).fetch('id')
       else
         log(:error, "Failure: #{response.class}", response, data)
+        nil
       end
-
     rescue => e
       log(:error, "[Honeybadger::Sender#send_to_honeybadger] Error: #{e.class} - #{e.message}\nBacktrace:\n#{e.backtrace.join("\n\t")}")
       nil
@@ -66,12 +65,9 @@ module Honeybadger
     def ping(data = {})
       return nil unless api_key_ok?
 
-      response = client.post do |p|
-        p.url "/v1/ping"
-        p.body = data.to_json
-      end
+      response = http_connection.post('/v1/ping', data.to_json, http_headers)
 
-      if response.success?
+      if Net::HTTPSuccess === response
         JSON.parse(response.body)
       else
         log(:error, "Ping Failure", response, data)
@@ -101,6 +97,10 @@ module Honeybadger
 
     private
 
+    def url
+      URI.parse("#{protocol}://#{host}:#{port}").merge(NOTICES_URI)
+    end
+
     def api_key_ok?(api_key = nil)
       api_key ||= self.api_key
       unless api_key =~ /\S/
@@ -111,38 +111,6 @@ module Honeybadger
       api_key
     end
 
-    def proxy_options
-      if proxy_host
-        { :uri => "#{protocol}://#{proxy_host}:#{proxy_port || port}", :user => proxy_user, :password => proxy_pass }
-      end
-    end
-
-    def request_options
-      { :timeout => http_read_timeout, :open_timeout => http_open_timeout }
-    end
-
-    def client
-      @client ||= Faraday.new(:request => request_options, :proxy => proxy_options).tap do |conn|
-        conn.build do |builder|
-          builder.adapter Faraday.default_adapter
-        end
-
-        conn.url_prefix = "#{protocol}://#{host}:#{port}"
-        conn.headers['User-agent'] = "HB-Ruby #{Honeybadger::VERSION}; #{RUBY_VERSION}; #{RUBY_PLATFORM}"
-        conn.headers['X-API-Key'] = api_key.to_s
-        conn.headers['Content-Type'] = 'application/json'
-        conn.headers['Accept'] = 'text/json, application/json'
-
-        if secure?
-          conn.ssl[:verify_mode] = OpenSSL::SSL::VERIFY_PEER
-          conn.ssl[:ca_file] = Honeybadger.configuration.ca_bundle_path
-        end
-      end
-    rescue => e
-      log(:error, "[Honeybadger::Sender#client] Failure initializing the HTTP connection.\nError: #{e.class} - #{e.message}\nBacktrace:\n#{e.backtrace.join("\n\t")}")
-      raise e
-    end
-
     def log(level, message, response = nil, data = nil)
       # Log result:
       Honeybadger.write_verbose_log(message, level)
@@ -151,6 +119,40 @@ module Honeybadger
       Honeybadger.report_environment_info
       Honeybadger.report_response_body(response.body) if response && response.body =~ /\S/
       Honeybadger.write_verbose_log("Notice: #{data}", :debug) if data && Honeybadger.configuration.debug
+    end
+
+    def http_connection
+      setup_http_connection
+    end
+
+    def http_headers(headers=nil)
+      {}.tap do |hash|
+        hash.merge!(HEADERS)
+        hash.merge!({'X-API-Key' => api_key})
+        hash.merge!(headers) if headers
+      end
+    end
+
+    def setup_http_connection
+      http_class = Net::HTTP::Proxy(proxy_host, proxy_port, proxy_user, proxy_pass)
+      http = http_class.new(url.host, url.port)
+
+      http.read_timeout = http_read_timeout
+      http.open_timeout = http_open_timeout
+
+      if secure?
+        http.use_ssl     = true
+
+        http.ca_file      = Honeybadger.configuration.ca_bundle_path
+        http.verify_mode  = OpenSSL::SSL::VERIFY_PEER
+      else
+        http.use_ssl     = false
+      end
+
+      http
+    rescue => e
+      log(:error, "[Honeybadger::Sender#setup_http_connection] Failure initializing the HTTP connection.\nError: #{e.class} - #{e.message}\nBacktrace:\n#{e.backtrace.join("\n\t")}")
+      raise e
     end
   end
 end