+++ spec/honeybadger/logger_spec.rb
@@ -28,7 +28,7 @@ describe Honeybadger do
   it "prints environment info on a failed notification without a body" do
     reset_config
     stub_verbose_log
-    stub_http(:response => Faraday::Response.new(:status => 500), :body => nil)
+    stub_http(:response => Net::HTTPError, :body => nil)
     Honeybadger.should_receive(:write_verbose_log).with(/Environment Info:/)
     Honeybadger.should_not_receive(:write_verbose_log).with(/Response from Honeybadger:/, anything)
     send_notice
@@ -46,7 +46,7 @@ describe Honeybadger do
   it "prints environment info and response on a failure with a body" do
     reset_config
     stub_verbose_log
-    stub_http(:response => Faraday::Response.new(:status => 500))
+    stub_http(:response => Net::HTTPError)
     Honeybadger.should_receive(:write_verbose_log).with(/Environment Info:/)
     Honeybadger.should_receive(:write_verbose_log).with(/Response from Honeybadger:/)
     send_notice