+++ spec/honeybadger/sender_spec.rb
@@ -12,18 +12,12 @@ describe Honeybadger::Sender do
   end
 
   it "posts to Honeybadger when using an HTTP proxy" do
-    post = double(:headers => {})
-    http = stub_http
-    http.stub(:post).and_yield(post).and_return(false)
+    http  = stub_http
+    proxy = double(:new => http)
+    Net::HTTP.stub(:Proxy).and_return(proxy)
 
-    url = "http://api.honeybadger.io:80#{Honeybadger::Sender::NOTICES_URI}"
-    uri = URI.parse(url)
-
-    post.should_receive(:url).with(uri.path)
-    post.should_receive(:body=).with('asdf')
-
-    Faraday.should_receive(:new).
-      with(hash_including(:proxy => { :uri => 'https://some.host:88', :user => 'login', :password => 'passwd' })).and_return(http)
+    http.should_receive(:post).with(Honeybadger::Sender::NOTICES_URI, kind_of(String), Honeybadger::HEADERS.merge({ 'X-API-Key' => 'abc123'}))
+    Net::HTTP.should_receive(:Proxy).with('some.host', 88, 'login', 'passwd')
 
     send_exception(:proxy_host => 'some.host',
                    :proxy_port => 88,
@@ -38,7 +32,7 @@ describe Honeybadger::Sender do
   end
 
   it "returns nil on failed posting" do
-    stub_http(:response => Faraday::Response.new(:status => 500))
+    stub_http(:response => Net::HTTPError)
     expect(send_exception(:secure => false)).to be_nil
   end
 
@@ -71,14 +65,14 @@ describe Honeybadger::Sender do
   it "logs success" do
     stub_http
     sender = build_sender
-    sender.should_receive(:log).with(:debug, /Success/, kind_of(Faraday::Response), kind_of(String))
+    sender.should_receive(:log).with(:debug, /Success/, kind_of(Net::HTTPSuccess), kind_of(String))
     send_exception(:sender => sender, :secure => false)
   end
 
   it "logs failure" do
-    stub_http(:response => Faraday::Response.new(:status => 500))
+    stub_http(:response => Net::HTTPServerError.new('1.2', '500', 'Internal Error'))
     sender = build_sender
-    sender.should_receive(:log).with(:error, /Failure/, kind_of(Faraday::Response), kind_of(String))
+    sender.should_receive(:log).with(:error, /Failure/, kind_of(Net::HTTPServerError), kind_of(String))
     send_exception(:sender => sender, :secure => false)
   end
 
@@ -86,24 +80,28 @@ describe Honeybadger::Sender do
     # TODO: Figure out why nested groups aren't running
     context "HTTP connection setup problems" do
       it "should not be rescued" do
-        Faraday.should_receive(:new).and_raise(NoMemoryError)
-        expect { build_sender.send(:client) }.to raise_error(NoMemoryError)
+        proxy = double()
+        proxy.stub(:new).and_raise(NoMemoryError)
+        Net::HTTP.stub(:Proxy).and_return(proxy)
+        expect { build_sender.send(:setup_http_connection) }.to raise_error(NoMemoryError)
       end
 
       it "should be logged" do
-        Faraday.should_receive(:new).and_raise(RuntimeError)
+        proxy = double()
+        proxy.stub(:new).and_raise(RuntimeError)
+        Net::HTTP.stub(:Proxy).and_return(proxy)
 
         sender = build_sender
         sender.should_receive(:log).with(:error, /Failure initializing the HTTP connection/)
 
-        expect { sender.send(:client) }.to raise_error(RuntimeError)
+        expect { sender.send(:setup_http_connection) }.to raise_error(RuntimeError)
       end
     end
 
     context "unexpected exception sending problems" do
       it "should be logged" do
         sender  = build_sender
-        sender.should_receive(:client).and_raise(RuntimeError)
+        sender.should_receive(:setup_http_connection).and_raise(RuntimeError)
 
         sender.should_receive(:log).with(:error, /Error/)
         sender.send_to_honeybadger("stuff")
@@ -111,7 +109,7 @@ describe Honeybadger::Sender do
 
       it "returns nil no matter what" do
         sender  = build_sender
-        sender.should_receive(:client).and_raise(LocalJumpError)
+        sender.should_receive(:setup_http_connection).and_raise(LocalJumpError)
 
         expect { sender.send_to_honeybadger("stuff").should be_nil }.not_to raise_error
       end
@@ -149,34 +147,32 @@ describe Honeybadger::Sender do
       http = stub_http
       url = "http://api.honeybadger.io:80#{Honeybadger::Sender::NOTICES_URI}"
       uri = URI.parse(url)
-      post = double(:body= => nil, :headers => {})
-      http.should_receive(:post).and_yield(post)
-      post.should_receive(:url).with(uri.path)
+      http.should_receive(:post).with(uri.path, anything, Honeybadger::HEADERS.merge({ 'X-API-Key' => 'abc123'}))
       send_exception(:secure => false)
     end
 
     it "post to the right path for ssl" do
       http = stub_http
-      post = double(:body= => nil, :headers => {})
-      http.should_receive(:post).and_yield(post)
-      post.should_receive(:url).with(Honeybadger::Sender::NOTICES_URI)
+      http.should_receive(:post).with(Honeybadger::Sender::NOTICES_URI, anything, Honeybadger::HEADERS.merge({ 'X-API-Key' => 'abc123'}))
       send_exception(:secure => true)
     end
 
     it "verifies the SSL peer when the use_ssl option is set to true" do
       url = "https://api.honeybadger.io#{Honeybadger::Sender::NOTICES_URI}"
+      uri = URI.parse(url)
 
-      real_http = Faraday.new(:url => url)
+      real_http = Net::HTTP.new(uri.host, uri.port)
       real_http.stub(:post => nil)
-      Faraday.stub(:new).and_return(real_http)
+      proxy = double(:new => real_http)
+      Net::HTTP.stub(:Proxy => proxy)
 
       File.stub(:exist?).with(OpenSSL::X509::DEFAULT_CERT_FILE).and_return(false)
 
       send_exception(:secure => true)
 
-      expect(real_http.ssl).not_to be_empty
-      expect(real_http.ssl[:verify_mode]).to eq OpenSSL::SSL::VERIFY_PEER
-      expect(real_http.ssl[:ca_file]).to eq Honeybadger.configuration.local_cert_path
+      expect(real_http.use_ssl?).to be_true
+      expect(real_http.verify_mode).to eq OpenSSL::SSL::VERIFY_PEER
+      expect(real_http.ca_file).to eq Honeybadger.configuration.local_cert_path
     end
 
     it "uses the default DEFAULT_CERT_FILE if asked to" do
@@ -191,61 +187,65 @@ describe Honeybadger::Sender do
 
       expect(sender.use_system_ssl_cert_chain?).to be_true
 
-      http    = sender.send(:client)
-      expect(http.ssl[:ca_file]).not_to eq Honeybadger.configuration.local_cert_path
+      http = sender.send(:setup_http_connection)
+      expect(http.ca_file).not_to eq Honeybadger.configuration.local_cert_path
     end
 
     it "verifies the connection when the use_ssl option is set (VERIFY_PEER)" do
       sender  = build_sender(:secure => true)
-      http    = sender.send(:client)
-      expect(http.ssl[:verify_mode]).to eq OpenSSL::SSL::VERIFY_PEER
+      http    = sender.send(:setup_http_connection)
+      expect(http.verify_mode).to eq OpenSSL::SSL::VERIFY_PEER
     end
 
     it "uses the default cert (OpenSSL::X509::DEFAULT_CERT_FILE) only if explicitly told to" do
       sender  = build_sender(:secure => true)
-      http    = sender.send(:client)
+      http    = sender.send(:setup_http_connection)
 
-      expect(http.ssl[:ca_file]).to eq Honeybadger.configuration.local_cert_path
+      expect(http.ca_file).to eq Honeybadger.configuration.local_cert_path
 
       File.stub(:exist?).with(OpenSSL::X509::DEFAULT_CERT_FILE).and_return(true)
       sender  = build_sender(:secure => true, :use_system_ssl_cert_chain => true)
-      http    = sender.send(:client)
+      http    = sender.send(:setup_http_connection)
 
-      expect(http.ssl[:ca_file]).not_to eq Honeybadger.configuration.local_cert_path
-      expect(http.ssl[:ca_file]).to eq OpenSSL::X509::DEFAULT_CERT_FILE
+      expect(http.ca_file).not_to eq Honeybadger.configuration.local_cert_path
+      expect(http.ca_file).to eq OpenSSL::X509::DEFAULT_CERT_FILE
     end
 
     it "uses ssl if secure" do
       sender  = build_sender(:secure => true)
-      http    = sender.send(:client)
+      http    = sender.send(:setup_http_connection)
       expect(http.port).to eq 443
     end
 
     it "does not use ssl if not secure" do
       sender  = build_sender(:secure => false)
-      http    = sender.send(:client)
+      http    = sender.send(:setup_http_connection)
       expect(http.port).to eq 80
     end
   end
 
   context "network timeouts" do
     it "default the open timeout to 2 seconds" do
-      Faraday.should_receive(:new).with(hash_including({ :request => hash_including({ :open_timeout => 2 }) }))
+      http = stub_http
+      http.should_receive(:open_timeout=).with(2)
       send_exception
     end
 
     it "default the read timeout to 5 seconds" do
-      Faraday.should_receive(:new).with(hash_including({ :request => hash_including({ :timeout => 5 }) }))
+      http = stub_http
+      http.should_receive(:read_timeout=).with(5)
       send_exception
     end
 
     it "allow override of the open timeout" do
-      Faraday.should_receive(:new).with(hash_including({ :request => hash_including({ :open_timeout => 4 }) }))
+      http = stub_http
+      http.should_receive(:open_timeout=).with(4)
       send_exception(:http_open_timeout => 4)
     end
 
     it "allow override of the read timeout" do
-      Faraday.should_receive(:new).with(hash_including({ :request => hash_including({ :timeout => 10 }) }))
+      http = stub_http
+      http.should_receive(:read_timeout=).with(10)
       send_exception(:http_read_timeout => 10)
     end
   end