+++ spec/support/helpers.rb
@@ -24,12 +24,15 @@ module Helpers
   end
 
   def stub_http(options = {})
-    response = options[:response] || Faraday::Response.new(:status => 200)
+    response = options[:response] || Net::HTTPSuccess.new('1.2', '200', 'OK')
     response.stub(:body => options[:body] || '{"id":"1234"}')
-
-    http = Faraday.new
-    http.stub(:post).and_return(response)
-    Faraday.stub(:new => http)
+    http = double(:post          => response,
+                :read_timeout= => nil,
+                :open_timeout= => nil,
+                :ca_file=      => nil,
+                :verify_mode=  => nil,
+                :use_ssl=      => nil)
+    Net::HTTP.stub(:new).and_return(http)
     http
   end